requirements.txt
requests
beautifulsoup4
openai
gunicorn
python-dotenv
scrapegraphai
playwright
numpy
gunicorn